Fix SD 1.5 NaN output: float32 compute precision and scheduler lower-order final
Two independent fixes for blank/NaN output from SD 1.5 and SD 2.1:

1. Change compute_precision from float16 to float32 for SD 1.5 and SD 2.1
   exports. Float16 attention Q*K^T products overflow fp16 max (65504)
   during later denoising steps, producing Inf then NaN after softmax.

2. Remove the timeSteps.count < 15 gate from DPM-Solver++ lowerOrderFinal.
   HuggingFace diffusers applies lower_order_final unconditionally. With
   20 steps, the second-order solver on the final step amplifies errors
   via 1/r0 (~5x), which compounds the fp16 overflow.

+++ python/src/coreai_models/model_registry.py
@@ -335,7 +335,7 @@ DIFFUSION_PRESETS: list[ModelPreset] = [
         "diffusion",
         None,
         "none",
-        "float16",
+        "float32",
         None,
     ),
     ModelPreset(
@@ -345,7 +345,7 @@ DIFFUSION_PRESETS: list[ModelPreset] = [
         "diffusion",
         None,
         "none",
-        "float16",
+        "float32",
         None,
     ),
     ModelPreset(
